requirements.txt
pycryptodome==3.10.1
numpy==1.20.1
prettytable